requirements.txt
scrapyrt==0.12
